Feat: getallrooms, getavailablerooms 관련 순서 문제 루트 페이지에서 api 호출하는 걸로 해결 완료
